backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
itsdangerous==2.2.0
passlib[argon2]==1.7.4
PyYAML==6.0.2
psutil==7.0.0
httpx==0.28.1
python-jose==3.5.0
websockets==15.0.1
pytest==8.4.1
